wojtek to debil

+++ Assets/Scripts/CoinsCounter.cs
@@ -7,6 +7,6 @@ public class CoinsCounter : MonoBehaviour
     public TMP_Text counter;
     void Update()
     {
-        counter.text = GameManager.Instance.money.ToString();
+        counter.text =  ((int)GameManager.Instance.money).ToString();
     }
 }
